move checksum model to the integrity pkg.
